+++ src/pygama/raw/orca/orca_flashcam.py
@@ -323,43 +323,3 @@ class ORFlashCamADCWaveformDecoder(ORFlashCamWaveformDecoder):
         self.ch_orca_shift = 10
         self.channel_mask = 0x000003FF
 
-
-class ORRunDecoderForRun(OrcaDecoder):
-    def __init__(self, header: OrcaHeader = None, **kwargs) -> None:
-        self.decoded_values = {
-            "subrun_number": {"dtype": "uint16"},
-            "runstartorstop": {"dtype": "bool8"},
-            "quickstartrun": {"dtype": "bool8"},
-            "remotecontrolrun": {"dtype": "bool8"},
-            "heartbeatrecord": {"dtype": "bool8"},
-            "endsubrunrecord": {"dtype": "bool8"},
-            "startsubrunrecord": {"dtype": "bool8"},
-            "run_number": {"dtype": "int32"},
-            "time": {"dtype": "int32"},
-        }
-        super().__init__(header=header, **kwargs)
-
-    def get_decoded_values(self, key: int = None) -> dict[str, Any]:
-        return self.decoded_values
-
-    def decode_packet(
-        self, packet: OrcaPacket, packet_id: int, rbl: RawBufferLibrary
-    ) -> bool:
-        if len(rbl) != 1:
-            log.warning(
-                f"got {len(rbl)} rb's, should have only 1 (no keyed decoded values)"
-            )
-        rb = rbl[0]
-        tbl = rb.lgdo
-        ii = rb.loc
-
-        tbl["subrun_number"].nda[ii] = (packet[1] & 0xFFFF0000) >> 16
-        for i, k in enumerate(self.decoded_values):
-            if 0 < i < 7:
-                tbl[k].nda[ii] = (packet[1] & (1 << (i - 1))) >> (i - 1)
-
-        tbl["run_number"].nda[ii] = packet[2]
-        tbl["time"].nda[ii] = packet[3]
-
-        rb.loc += 1
-        return rb.is_full()
