+++ src/pygama/raw/orca/orca_streamer.py
@@ -17,9 +17,9 @@ from pygama.raw.orca.orca_flashcam import (  # noqa: F401
     ORFlashCamADCWaveformDecoder,
     ORFlashCamListenerConfigDecoder,
     ORFlashCamWaveformDecoder,
-    ORRunDecoderForRun,
 )
 from pygama.raw.orca.orca_header_decoder import OrcaHeaderDecoder
+from pygama.raw.orca.orca_run_decoder import ORRunDecoderForRun
 from pygama.raw.raw_buffer import RawBuffer, RawBufferLibrary
 
 log = logging.getLogger(__name__)
